!CONFIG CHANGE!
climate path now must be specified as fullpath. climate directory no longer defaults to proj_dir + inputs/climate.

This enable climate data to reside anywhere - important if it is big and different simulations depend upon it - avoids having to hold replicates on disk.

This is especially for the SLURM scheduler use case Joel is developing (and others)